refactor: add SearchInput add SearchResult
