+++ src/Interfaces.ts
@@ -1,4 +1,4 @@
-import { ReactNode } from 'react';
+import { ChangeEvent, ReactNode } from 'react';
 
 export interface Character {
   name: string;
@@ -29,4 +29,19 @@ export interface ErrorBoundaryProps {
 export interface ErrorBoundaryState {
   hasError: boolean;
   errorMessage: string;
+}
+
+export interface SearchInputProps {
+  searchTerm: string;
+  onInputChange: (event: ChangeEvent<HTMLInputElement>) => void;
+  onSearch: () => void;
+  onErrorTest: () => void;
+}
+
+export interface SearchResultProps {
+  loading: boolean;
+  error: boolean;
+  errorMessage: string;
+  searchTerm: string;
+  searchData: Character[] | null;
 }
