+++ src/components/SearchResult/SearchResult.css
@@ -1,54 +1,3 @@
-body {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 0;
-}
-
-.content {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: center;
-  max-width: 1400px;
-  padding: 20px;
-}
-
-.search_side {
-  background-color: #2a2a40;
-  width: 100%;
-  max-width: 800px;
-  text-align: center;
-  padding: 20px;
-  border-radius: 8px;
-  margin-bottom: 20px;
-}
-
-.search_side input {
-  padding: 10px;
-  font-size: 16px;
-  border: 1px solid #444;
-  border-radius: 4px;
-  margin-right: 10px;
-  background-color: #333;
-  color: #d1d1e0;
-}
-
-.search_side button {
-  padding: 10px 20px;
-  font-size: 16px;
-  border: none;
-  border-radius: 4px;
-  background-color: #434355;
-  color: white;
-  cursor: pointer;
-  margin: 10px;
-}
-
-.search_side button:hover {
-  background-color: #6d6d74;
-}
-
 .result_side {
   width: 100%;
   max-width: 800px;
